Movimiento rana acorde a filas y columnas arreglado

+++ Backend/frog.c
@@ -1,7 +1,7 @@
 #include "frog.h"
 
 // Función para inicializar la rana con valores predeterminados
-void init_frog(frog_t *frog, int8_t x, int8_t y, uint8_t vidas, int16_t puntos, int8_t llegadas) {
+void init_frog(frog_t *frog, float x, float y, uint8_t vidas, int16_t puntos, int8_t llegadas) {
     frog->x = x;
     frog->y = y;
     frog->vidas = vidas;
@@ -10,11 +10,11 @@ void init_frog(frog_t *frog, int8_t x, int8_t y, uint8_t vidas, int16_t puntos, 
 }
 
 // Funciones para obtener los valores
-int8_t get_frog_x(const frog_t *frog) {
+float get_frog_x(const frog_t *frog) {
     return frog->x;
 }
 
-int8_t get_frog_y(const frog_t *frog) {
+float get_frog_y(const frog_t *frog) {
     return frog->y;
 }
 
@@ -31,11 +31,11 @@ int8_t get_frog_llegadas(const frog_t *frog) {
 }
 
 // Funciones para modificar los valores
-void set_frog_x(frog_t *frog, int8_t x) {
+void set_frog_x(frog_t *frog, float x) {
     frog->x = x;
 }
 
-void set_frog_y(frog_t *frog, int8_t y) {
+void set_frog_y(frog_t *frog, float y) {
     frog->y = y;
 }
 
@@ -69,7 +69,7 @@ void increase_frog_llegadas(frog_t *frog) {
 }
 
 // Función para mover la rana
-void move_frog(frog_t *frog, int8_t dx, int8_t dy) {
+void move_frog(frog_t *frog, float dx, float dy) {
     frog->x += dx;
     frog->y += dy;
 }
